Implement client and server modes
