+++ Cargo.toml
@@ -10,3 +10,4 @@ aes-gcm = { version = "0.10", features = ["aes"] }
 rand = "0.8"
 sha2 = "0.10"
 tun = "0.6"
+clap = { version = "4", features = ["derive"] }
