+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod pqc;
 pub mod crypto;
 pub mod ipv6;
 pub mod tundev;
+pub mod modes;
